Move AutosysSensor / Code cleanup
