+++ brickflow_plugins/databricks/external_tasks.py
@@ -364,82 +364,3 @@ class TaskDependencySensor(BaseSensorOperator):
             if (time.time() - self._start_time) > self.timeout:
                 raise AirflowSensorTimeout("The job has timed out")
         log.info(f"Upstream Dag {external_dag_id} is successful")
-
-
-class AutosysSensor(BaseSensorOperator):
-    def __init__(
-        self,
-        url: str,
-        job_name: str,
-        poke_interval: int,
-        time_delta: Union[timedelta, dict] = {"days": 0},
-        *args,
-        **kwargs,
-    ):
-        super().__init__(*args, **kwargs)
-        self.url = url
-        self.job_name = job_name
-        self.poke_interval = poke_interval
-        self.time_delta = time_delta
-        self.url = self.url + self.job_name
-
-    """
-        Takes in url, job_name, poke_interval, execution delta and airflow_cluster_auth as parameters 
-        and sends a http get() request, checks the API response and exits the process 
-        if the specified conditions are met.
-        If not, waits for the given poke interval, then pokes again and again until the conditions
-        are met or times out.
-    """
-
-    def poke(self, context):
-        logging.info("Poking: " + self.url)
-
-        headers = {
-            "Accept": "application/json",
-            "cache-control": "no-cache",
-        }
-
-        response = requests.get(
-            self.url,
-            headers=headers,
-            verify=False,  # nosec
-            timeout=10,
-        )
-
-        if response.status_code != 200:
-            raise HTTPError(
-                f"Request failed with '{response.status_code}' code. \n{response.text}"
-            )
-        else:
-            status = response.json()["status"][:2].upper()
-
-            last_end_timestamp = None
-            if last_end_utc := response.json().get("lastEndUTC"):
-                last_end_timestamp = parse(last_end_utc).replace(tzinfo=pytz.UTC)
-
-            time_delta = (
-                self.time_delta
-                if isinstance(self.time_delta, timedelta)
-                else timedelta(**self.time_delta)
-            )
-
-            execution_timestamp = parse(context["execution_date"])
-            run_timestamp = execution_timestamp - time_delta
-
-            if (
-                "SU" in status
-                and last_end_timestamp
-                and last_end_timestamp >= run_timestamp
-            ):
-                logging.info(
-                    f"Last End: {last_end_timestamp}, Run Timestamp: {run_timestamp}"
-                )
-                logging.info("Success criteria met. Exiting")
-                return True
-            else:
-                logging.info(
-                    f"Last End: {last_end_timestamp}, Run Timestamp: {run_timestamp}"
-                )
-                time.sleep(self.poke_interval)
-                logging.info("Poking again")
-                AutosysSensor.poke(self, context)
